Revert "Tim/expired connector"
