+++ package.json
@@ -65,7 +65,6 @@
     "octokit": "^5.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^7.9.5",
     "react-syntax-highlighter": "^15.6.1",
     "react-use": "^17.6.0",
     "remark-gfm": "^4.0.1",
@@ -100,7 +99,6 @@
     "@types/node": "^20.19.22",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "@types/react-router-dom": "^5.3.3",
     "@types/react-syntax-highlighter": "^15.5.13",
     "@types/ws": "^8.5.9",
     "@vitejs/plugin-react": "^4.0.0",
