+++ src/per-provider-token-storage.ts
@@ -146,16 +146,12 @@ export class PerProviderTokenStorage {
         file.startsWith('oauth-tokens.') && file.endsWith('.encrypted'),
       )
 
-      // Load all provider tokens in parallel
-      await Promise.all(
-        tokenFiles.map((file) => {
-          const providerId = file.replace('oauth-tokens.', '').replace('.encrypted', '')
-          if (!this.loadedProviders.has(providerId)) {
-            return this.loadProviderTokens(providerId)
-          }
-          return Promise.resolve()
-        }),
-      )
+      for (const file of tokenFiles) {
+        const providerId = file.replace('oauth-tokens.', '').replace('.encrypted', '')
+        if (!this.loadedProviders.has(providerId)) {
+          await this.loadProviderTokens(providerId)
+        }
+      }
     }
     catch (error) {
       console.error('❌ Error loading all provider tokens:', error)
@@ -267,25 +263,24 @@ export class PerProviderTokenStorage {
   }>> {
     await this.loadAllProviderTokens()
 
-    // Process all providers in parallel
-    const statusEntries = await Promise.all(
-      Array.from(this.tokensCache.entries()).map(async ([providerId, tokens]) => {
-        const expired = await this.areTokensExpired(providerId)
-        return [
-          providerId,
-          {
-            authenticated: true,
-            expired,
-            user: tokens.user,
-            storedAt: tokens.storedAt,
-            updatedAt: tokens.updatedAt,
-          },
-        ] as const
-      }),
-    )
+    const status: Record<string, {
+      authenticated: boolean
+      expired: boolean
+      user?: UserInfo
+      storedAt?: number
+      updatedAt?: number
+    }> = {}
 
-    // Convert array of entries back to an object
-    const status = Object.fromEntries(statusEntries)
+    for (const [providerId, tokens] of this.tokensCache.entries()) {
+      const expired = await this.areTokensExpired(providerId)
+      status[providerId] = {
+        authenticated: true,
+        expired,
+        user: tokens.user,
+        storedAt: tokens.storedAt,
+        updatedAt: tokens.updatedAt,
+      }
+    }
 
     return status
   }
@@ -419,39 +414,6 @@ export class PerProviderTokenStorage {
     }
     catch (error) {
       console.error('Error clearing onboarding token:', error)
-      throw error
-    }
-  }
-
-  /**
-   * Expire all tokens for testing auto-refresh functionality
-   * Sets expires_at to a past timestamp without invalidating refresh tokens
-   */
-  async expireAllTokensForTesting(): Promise<number> {
-    try {
-      // Load all provider tokens
-      await this.loadAllProviderTokens()
-
-      let expiredCount = 0
-
-      // Expire each token
-      for (const [providerId, tokens] of this.tokensCache.entries()) {
-        // Set expires_at to 1 hour ago
-        tokens.expires_at = Date.now() - 3600000
-        tokens.updatedAt = Date.now()
-
-        // Save the updated tokens
-        await this.saveProviderTokens(tokens)
-        expiredCount++
-
-        console.log(`🧪 Expired tokens for ${providerId} for testing`)
-      }
-
-      console.log(`🧪 Expired ${expiredCount} provider token(s) for testing`)
-      return expiredCount
-    }
-    catch (error) {
-      console.error('❌ Error expiring tokens for testing:', error)
       throw error
     }
   }
