+++ src/preload.ts
@@ -118,9 +118,8 @@ export interface ProgressInfo {
 }
 
 export interface ElectronAPI {
-  sendMessageResponse: (message: Message, feedback?: string) => Promise<void>
-  approveMessage: (message: Message, feedback?: string) => Promise<void> // @deprecated Use sendMessageResponse instead
-  rejectMessage: (messageId: string, feedback?: string) => Promise<void> // @deprecated Use sendMessageResponse instead
+  approveMessage: (message: Message, feedback?: string) => Promise<void>
+  rejectMessage: (messageId: string, feedback?: string) => Promise<void>
   approveCollectionShare: (messageId: string, updatedRequest: CollectionRequest) => Promise<void>
   rejectCollectionShare: (messageId: string) => Promise<void>
   sendPromptCollectionRequest: (context: { scripts: Script[], prompt: string, images: string[] }) => Promise<string>
@@ -165,7 +164,6 @@ export interface ElectronAPI {
   getProviderTokens: (providerId: string) => Promise<ProviderTokens>
   refreshProviderTokens: (providerId: string) => Promise<boolean>
   clearAllProviderTokens: () => Promise<void>
-  expireAllTokensForTesting: () => Promise<number>
   getOAuthStorageInfo: () => Promise<OAuthStorageInfo>
   onProviderAuthSuccess: (callback: (event: IpcRendererEvent, data: ProviderAuthEventData) => void) => void
   onProviderAuthError: (callback: (event: IpcRendererEvent, error: ProviderAuthErrorData) => void) => void
@@ -254,14 +252,7 @@ export interface ElectronAPI {
 // Expose protected methods that allow the renderer process to use
 // the ipcRenderer without exposing the entire object
 contextBridge.exposeInMainWorld('electronAPI', {
-  sendMessageResponse: (message: Message, feedback?: string): Promise<void> => {
-    return ipcRenderer.invoke('send-message-response', message, feedback)
-  },
-  // @deprecated - kept for backward compatibility
-  approveMessage: (message: Message, feedback?: string): Promise<void> => {
-    return ipcRenderer.invoke('approve-message', message, feedback)
-  },
-  // @deprecated - kept for backward compatibility
+  approveMessage: (message: Message, feedback?: string): Promise<void> => ipcRenderer.invoke('approve-message', message, feedback),
   rejectMessage: (messageId: string, feedback?: string): Promise<void> => ipcRenderer.invoke('reject-message', messageId, feedback),
   approveCollectionShare: (messageId: string, updatedRequest: CollectionRequest): Promise<void> => ipcRenderer.invoke('approve-collection-share', messageId, updatedRequest),
   rejectCollectionShare: (messageId: string): Promise<void> => ipcRenderer.invoke('reject-collection-share', messageId),
@@ -352,7 +343,6 @@ contextBridge.exposeInMainWorld('electronAPI', {
   getProviderTokens: (providerId: string): Promise<ProviderTokens> => ipcRenderer.invoke('get-provider-tokens', providerId),
   refreshProviderTokens: (providerId: string): Promise<boolean> => ipcRenderer.invoke('refresh-provider-tokens', providerId),
   clearAllProviderTokens: (): Promise<void> => ipcRenderer.invoke('clear-all-provider-tokens'),
-  expireAllTokensForTesting: (): Promise<number> => ipcRenderer.invoke('expire-all-tokens-for-testing'),
   getOAuthStorageInfo: (): Promise<OAuthStorageInfo> => ipcRenderer.invoke('get-oauth-storage-info'),
 
   // OAuth Provider event listeners
