+++ src/renderer/components/ui/ButtonDesigned.tsx
@@ -1,19 +1,18 @@
 import React from 'react'
 import { cn } from '../../lib/utils'
 
-const CLEAR_BUTTON_CLASS = 'bg-transparent hover:bg-transparent active:bg-transparent disabled:bg-transparent disabled:text-[#737373]'
-const DESTRUCTIVE_BUTTON_CLASS = 'text-[#D23535] hover:text-[#FFF] bg-[#F7F7F7] hover:bg-[#D23535] disabled:bg-[#F7F7F7] disabled:text-[#D9D9D9]'
-const PRIMARY_BLACK_BUTTON_CLASS = 'text-white bg-[#171717] hover:bg-[#171717] active:bg-[#171717] disabled:bg-[#171717] disabled:text-white'
 const PRIMARY_BUTTON_CLASS = 'text-[#5093B7] bg-[#E4EBEF] hover:bg-[#D5E0E6] active:bg-[#C5D4DD] disabled:bg-[#E4EBEF] disabled:text-[#5093B7]'
 const SECONDARY_BUTTON_CLASS = 'text-[#737373] bg-[#F3F3F3] hover:bg-[#E6E6E6] active:bg-[#D9D9D9] disabled:bg-[#F3F3F3] disabled:text-[#D9D9D9]'
+const DESTRUCTIVE_BUTTON_CLASS = 'text-[#D23535] hover:text-[#FFF] bg-[#F7F7F7] hover:bg-[#D23535] disabled:bg-[#F7F7F7] disabled:text-[#D9D9D9]'
+const CLEAR_BUTTON_CLASS = 'bg-transparent hover:bg-transparent active:bg-transparent disabled:bg-transparent disabled:text-[#737373]'
 
 export const ButtonDesigned: React.FC<{
   children: React.ReactNode
   className?: string
   disabled?: boolean
   hasBorder?: boolean
-  onClick?: (event: React.MouseEvent<HTMLButtonElement>) => void
-  variant: 'primary' | 'primary-black' | 'secondary' | 'destructive' | 'clear'
+  onClick: () => void
+  variant: 'primary' | 'secondary' | 'destructive' | 'clear'
 }> = ({
   children,
   className,
@@ -26,9 +25,6 @@ export const ButtonDesigned: React.FC<{
   switch (variant) {
     case 'primary':
       variantClass = PRIMARY_BUTTON_CLASS
-      break
-    case 'primary-black':
-      variantClass = PRIMARY_BLACK_BUTTON_CLASS
       break
     case 'secondary':
       variantClass = SECONDARY_BUTTON_CLASS
@@ -45,7 +41,7 @@ export const ButtonDesigned: React.FC<{
   return (
     <button
       className={cn(
-        'inline-flex gap-[0.31rem] rounded-[0.25rem] p-[0.5rem] items-center justify-center disabled:opacity-50 disabled:cursor-not-allowed',
+        'gap-[0.31rem] rounded-[0.25rem] p-[0.5rem] items-center justify-center disabled:opacity-50 disabled:cursor-not-allowed',
         variantClass,
         borderClass,
         className,
