--- src/services/oauth-service.ts
@@ -1,844 +1,0 @@
-import * as crypto from 'crypto'
-import { shell } from 'electron'
-import { CodespaceEncryptionConfig, encryptWithCodespaceKey } from '../codespace-encryption'
-import { GithubService } from '../Github'
-import { OAuthCallbackData, OAuthHttpServer } from '../oauth-http-server'
-import { PKCEParams as NewPKCEParams, OAuthProvider, OAuthProviderManager, ProviderTokens, ServerProvider, ServerProviderInfo } from '../oauth-providers'
-import { OAuthTokenStorage, StoredProviderTokens } from '../oauth-token-storage'
-import { PerProviderTokenStorage } from '../per-provider-token-storage'
-import { OAuthProviderConfig } from '../provider-storage'
-import { Message } from '../types'
-import { ExecutorWebSocketClient } from '../websocket-client-to-executor'
-import { WindowManager } from '../window-manager'
-
-interface OnboardingGitHubResponse {
-  session_id: string
-  authorization_url: string
-  state: string
-}
-
-/**
- * OAuthService handles third-party provider OAuth flows (Google, GitHub, Slack, etc.)
- * Responsibilities:
- * - Provider OAuth initialization and management
- * - Direct provider OAuth flows
- * - Server-based provider OAuth flows
- * - Token refresh for providers
- * - Provider token encryption
- * - Executor WebSocket token requests
- * - Provider status management
- */
-export class OAuthService {
-  private oauthProviderManager!: OAuthProviderManager
-  private oauthTokenStorage!: OAuthTokenStorage
-  private perProviderTokenStorage!: PerProviderTokenStorage
-  private currentProviderPKCE: NewPKCEParams | null = null
-  private oauthHttpServer: OAuthHttpServer
-
-  constructor(
-    private windowManager: WindowManager,
-    private showNotification: (message: Message) => void,
-    private getExecutorWSClient: () => ExecutorWebSocketClient | null,
-    private getGithubService: () => GithubService,
-    private getEncryptionKey: () => string | null,
-    private getMainAccessToken: () => Promise<string | null>,
-    private readonly OAUTH_PORT: number,
-    private readonly SKIP_AUTH: boolean,
-  ) {
-    this.oauthHttpServer = new OAuthHttpServer(this.OAUTH_PORT)
-  }
-
-  /**
-   * Initialize OAuth provider system after encryption is ready
-   */
-  async initializeOAuthProviderSystem(): Promise<void> {
-    try {
-      // Initialize OAuth provider system
-      this.oauthProviderManager = new OAuthProviderManager()
-      this.oauthTokenStorage = new OAuthTokenStorage() // Keep for migration
-      this.perProviderTokenStorage = new PerProviderTokenStorage()
-
-      // Inject main access token getter for server provider refresh
-      this.oauthProviderManager.setMainAccessTokenGetter(() => this.getMainAccessToken())
-
-      // Migrate from old storage format
-      await this.migrateTokenStorage()
-
-      // Automatically refresh expired providers on startup
-      await this.refreshAllExpiredProvidersOnStartup()
-    }
-    catch (error) {
-      console.error('❌ Failed to initialize OAuth provider system:', error)
-      throw error
-    }
-  }
-
-  /**
-   * Refresh all expired OAuth providers on app startup
-   * This runs once during initialization to silently refresh tokens before the UI loads
-   */
-  private async refreshAllExpiredProvidersOnStartup(): Promise<void> {
-    try {
-      // Get all provider statuses
-      const providerStatuses = await this.perProviderTokenStorage.getProviderStatus()
-
-      // Find expired providers (authenticated but expired)
-      const expiredProviders = Object.entries(providerStatuses)
-        .filter(([, status]) => status?.authenticated && status?.expired)
-        .map(([providerId]) => providerId)
-
-      if (expiredProviders.length === 0) {
-        console.log('✅ No expired OAuth providers to refresh on startup')
-        return
-      }
-
-      console.log(`🔄 Refreshing ${expiredProviders.length} expired OAuth provider(s) on startup:`, expiredProviders)
-
-      let successCount = 0
-      let failedCount = 0
-
-      // Attempt to refresh each expired provider
-      // Do this sequentially to avoid rate limits and be gentle on startup
-      for (const providerId of expiredProviders) {
-        try {
-          const tokens = await this.perProviderTokenStorage.getTokens(providerId)
-          console.log('🔄 Refreshing tokens for', providerId, tokens)
-          if (!tokens?.refresh_token) {
-            console.log(`⚠️ Skipping ${providerId}: no refresh token available`)
-            failedCount++
-            continue
-          }
-
-          // Attempt to refresh the tokens
-          const refreshedTokens = await this.refreshProviderTokens(providerId, tokens.refresh_token)
-          await this.perProviderTokenStorage.storeTokens(refreshedTokens)
-
-          console.log(`✅ Successfully refreshed ${providerId} on startup`)
-          successCount++
-        }
-        catch (error) {
-          // Silently fail - user can manually refresh from the UI if needed
-          console.log(`⚠️ Failed to refresh ${providerId} on startup (will require manual refresh):`, error instanceof Error ? error.message : 'Unknown error')
-          failedCount++
-        }
-      }
-
-      console.log(`🔄 Startup OAuth refresh complete: ${successCount} succeeded, ${failedCount} failed`)
-    }
-    catch (error) {
-      // Don't throw - we don't want startup refresh failures to break app initialization
-      console.error('❌ Error during startup OAuth provider refresh:', error)
-    }
-  }
-
-  /**
-   * Migrate tokens from old single-file storage to new per-provider storage
-   */
-  private async migrateTokenStorage(): Promise<void> {
-    try {
-      const oldTokens = await this.oauthTokenStorage.getAllTokens()
-      if (Object.keys(oldTokens).length > 0) {
-        await this.perProviderTokenStorage.migrateFromOldStorage(oldTokens)
-
-        // After successful migration, optionally clear old storage
-        // Uncomment the next line if you want to remove the old file after migration
-        // await this.oauthTokenStorage.clearAllTokens();
-      }
-    }
-    catch (error) {
-      console.error('❌ Error during token storage migration:', error)
-    }
-  }
-
-  /**
-   * Helper function to encrypt provider token using codespace encryption
-   */
-  private async encryptProviderToken(token: string): Promise<{
-    encryptedToken: string
-    encrypted: boolean
-    encryptionMethod: string
-  }> {
-    let encryptedToken = token
-    let encrypted = false
-    let encryptionMethod = 'none'
-
-    if (token) {
-      try {
-        const onboardingToken = await this.perProviderTokenStorage.getValidAccessToken(
-          'onboarding',
-          this.refreshProviderTokens.bind(this),
-        )
-        const accessToken = await this.getMainAccessToken()
-
-        if (onboardingToken && accessToken) {
-          const config: CodespaceEncryptionConfig = {
-            codespaceUrl: 'auto', // Will be auto-discovered
-            bearerToken: accessToken, // Use GitHub token for authentication
-            githubToken: onboardingToken,
-          }
-
-          // This will auto-discover the codespace URL and encrypt the token
-          encryptedToken = await encryptWithCodespaceKey(token, config)
-          encrypted = true
-          encryptionMethod = 'rsa-codespace'
-        }
-      }
-      catch (encryptionError) {
-        console.error('❌ Failed to encrypt provider token:', encryptionError)
-        // Continue with unencrypted token as fallback
-      }
-    }
-
-    return { encryptedToken, encrypted, encryptionMethod }
-  }
-
-  /**
-   * Handle provider token request from executor WebSocket
-   */
-  async handleExecutorProviderTokenRequest(message: { providerId?: string, requestId?: string }): Promise<void> {
-    const { providerId } = message
-    const executorWSClient = this.getExecutorWSClient()
-
-    if (!providerId) {
-      // Send error response back through executor client
-      if (executorWSClient) {
-        executorWSClient.send({
-          type: 'provider-auth-token',
-          error: 'Provider ID is required',
-          timestamp: Date.now(),
-          requestId: message.requestId,
-        })
-      }
-      return
-    }
-
-    try {
-      const token = await this.getValidProviderAccessToken(providerId.toLowerCase())
-      const providerStatus = await this.perProviderTokenStorage.getProviderStatus()
-      const providerInfo = providerStatus[providerId]
-      const provider = await this.oauthProviderManager.getProvider(providerId)
-      if (!token) {
-        throw new Error('No token available')
-      }
-      // Encrypt token using codespace encryption if available
-
-      const { encryptedToken, encrypted, encryptionMethod } = await this.encryptProviderToken(token)
-      if (!encrypted) {
-        throw new Error('Failed to encrypt token')
-      }
-
-      const tokenResponse = {
-        type: 'provider-auth-token',
-        providerId: providerId,
-        token: encryptedToken,
-        encrypted: encrypted,
-        encryptionMethod: encryptionMethod,
-        timestamp: Date.now(),
-        requestId: message.requestId,
-        authenticated: !!token || this.SKIP_AUTH,
-        user: providerInfo?.user || (this.SKIP_AUTH ? { email: 'test@example.com', firstName: 'Test Provider' } : null),
-        providerName: provider?.name || providerId,
-      }
-
-      // Send response back through executor client
-      if (executorWSClient) {
-        executorWSClient.send(tokenResponse)
-      }
-    }
-    catch (error) {
-      // Send error response back through executor client
-      if (executorWSClient) {
-        executorWSClient.send({
-          type: 'provider-auth-token',
-          providerId: providerId,
-          error: `Failed to get token: ${error instanceof Error ? error.message : 'Unknown error'}`,
-          timestamp: Date.now(),
-          requestId: message.requestId,
-        })
-      }
-    }
-  }
-
-  /**
-   * Handle provider status request from executor WebSocket
-   */
-  async handleExecutorProviderStatusRequest(message: { requestId?: string }): Promise<void> {
-    const executorWSClient = this.getExecutorWSClient()
-
-    try {
-      const providerStatus = await this.perProviderTokenStorage.getProviderStatus()
-
-      // Check ALL stored provider tokens (both direct and server provider tokens)
-      const tokensAvailable = Object.entries(providerStatus)
-        .filter(([, status]) => status?.authenticated)
-        .map(([providerId]) => `KEYBOARD_PROVIDER_USER_TOKEN_FOR_${providerId.toUpperCase()}`)
-
-      const statusResponse = {
-        type: 'user-tokens-available',
-        tokensAvailable: tokensAvailable,
-        timestamp: Date.now(),
-        requestId: message.requestId,
-      }
-
-      // Send response back through executor client
-      if (executorWSClient) {
-        executorWSClient.send(statusResponse)
-      }
-    }
-    catch (error) {
-      // Send error response back through executor client
-      if (executorWSClient) {
-        executorWSClient.send({
-          type: 'user-tokens-available',
-          error: `Failed to get provider status: ${error instanceof Error ? error.message : 'Unknown error'}`,
-          timestamp: Date.now(),
-          requestId: message.requestId,
-        })
-      }
-    }
-  }
-
-  /**
-   * Start OAuth flow for a provider
-   */
-  async startProviderOAuthFlow(providerId: string): Promise<void> {
-    try {
-      const provider = await this.oauthProviderManager.getProvider(providerId)
-      if (!provider) {
-        throw new Error(`Provider ${providerId} not found`)
-      }
-
-      if (!provider.clientId) {
-        throw new Error(`Provider ${providerId} is not configured (missing client ID)`)
-      }
-
-      // Generate PKCE parameters
-      this.currentProviderPKCE = this.oauthProviderManager.generatePKCE(providerId)
-
-      // Start HTTP server to handle OAuth callback
-      await this.oauthHttpServer.startServer((callbackData: OAuthCallbackData) => {
-        this.handleOAuthHttpCallback(callbackData)
-      })
-
-      // Build authorization URL
-      const authUrl = await this.oauthProviderManager.buildAuthorizationUrl(providerId, this.currentProviderPKCE)
-
-      // Open browser for user authentication
-      await shell.openExternal(authUrl)
-    }
-    catch (error) {
-      console.error(`❌ OAuth flow error for ${providerId}:`, error)
-      await this.notifyProviderAuthError(providerId, 'Failed to start authentication')
-      this.oauthHttpServer.stopServer() // Clean up on error
-      throw error
-    }
-  }
-
-  /**
-   * Handle OAuth callback from HTTP server
-   */
-  private async handleOAuthHttpCallback(callbackData: OAuthCallbackData): Promise<void> {
-    try {
-      if (callbackData.error) {
-        throw new Error(`OAuth error: ${callbackData.error} - ${callbackData.error_description || ''}`)
-      }
-
-      if (!callbackData.code || !callbackData.state) {
-        throw new Error('Missing authorization code or state')
-      }
-
-      if (!this.currentProviderPKCE) {
-        throw new Error('No PKCE parameters stored - possible callback timeout or duplicate callback')
-      }
-
-      if (callbackData.state !== this.currentProviderPKCE.state) {
-        console.error('❌ State mismatch details:', {
-          received: callbackData.state,
-          expected: this.currentProviderPKCE.state,
-          providerId: this.currentProviderPKCE.providerId,
-        })
-        throw new Error('State mismatch - potential CSRF attack')
-      }
-
-      // Exchange code for tokens
-      await this.exchangeProviderCodeForTokens(this.currentProviderPKCE.providerId, callbackData.code, this.currentProviderPKCE)
-    }
-    catch (error) {
-      console.error('❌ OAuth HTTP callback error:', error)
-      const providerId = this.currentProviderPKCE?.providerId || 'unknown'
-      this.notifyProviderAuthError(providerId, `Authentication failed: ${error instanceof Error ? error.message : 'Unknown error'}`)
-    }
-  }
-
-  /**
-   * Exchange authorization code for provider tokens
-   */
-  private async exchangeProviderCodeForTokens(providerId: string, code: string, pkceParams: NewPKCEParams): Promise<void> {
-    try {
-      // Exchange code for tokens using provider manager
-      const tokens = await this.oauthProviderManager.exchangeCodeForTokens(providerId, code, pkceParams)
-
-      // Store tokens securely
-      await this.perProviderTokenStorage.storeTokens(tokens)
-
-      // Clear PKCE data
-      this.currentProviderPKCE = null
-
-      // Get provider info for notifications
-      const provider = await this.oauthProviderManager.getProvider(providerId)
-      const providerName = provider?.name || providerId
-
-      // Notify the renderer process
-      this.windowManager.sendMessage('provider-auth-success', {
-        providerId: providerId,
-        providerName: providerName,
-        user: tokens.user,
-        authenticated: true,
-      })
-
-      // Show the window after successful authentication
-      this.windowManager.showWindow()
-
-      // Show success notification
-      this.showNotification({
-        id: `auth-success-${providerId}`,
-        title: `${providerName} Authentication Successful`,
-        body: `Successfully connected to ${providerName}${tokens.user ? ` as ${tokens.user.name || tokens.user.email}` : ''}`,
-        timestamp: Date.now(),
-        priority: 'normal',
-      })
-    }
-    catch (error) {
-      console.error(`❌ Token exchange error for ${providerId}:`, error)
-      await this.notifyProviderAuthError(providerId, `Token exchange failed: ${error instanceof Error ? error.message : 'Unknown error'}`)
-    }
-  }
-
-  /**
-   * Refresh provider tokens
-   */
-  async refreshProviderTokens(providerId: string, refreshToken: string): Promise<ProviderTokens> {
-    return await this.oauthProviderManager.refreshTokens(providerId, refreshToken)
-  }
-
-  /**
-   * Get valid provider access token, refreshing if necessary
-   */
-  async getValidProviderAccessToken(providerId: string): Promise<string | null> {
-    return await this.perProviderTokenStorage.getValidAccessToken(
-      providerId,
-      this.refreshProviderTokens.bind(this),
-    )
-  }
-
-  /**
-   * Notify about provider authentication errors
-   */
-  private async notifyProviderAuthError(providerId: string, message: string): Promise<void> {
-    console.error(`🔐 Auth Error for ${providerId}:`, message)
-
-    this.windowManager.sendMessage('provider-auth-error', {
-      providerId,
-      message,
-    })
-
-    const provider = await this.oauthProviderManager.getProvider(providerId)
-    const providerName = provider?.name || providerId
-
-    this.showNotification({
-      id: `auth-error-${providerId}`,
-      title: `${providerName} Authentication Error`,
-      body: message,
-      timestamp: Date.now(),
-      priority: 'high',
-    })
-  }
-
-  /**
-   * Logout from a provider
-   */
-  async logoutProvider(providerId: string): Promise<void> {
-    await this.perProviderTokenStorage.removeTokens(providerId)
-    this.windowManager.sendMessage('provider-auth-logout', { providerId })
-  }
-
-  /**
-   * Start server provider OAuth flow
-   */
-  async startServerProviderOAuthFlow(serverId: string, provider: string): Promise<void> {
-    try {
-      const server = await this.oauthProviderManager.getServerProvider(serverId)
-      if (!server) {
-        throw new Error(`Server provider ${serverId} not found`)
-      }
-
-      // Generate state for the flow
-      const state = crypto.randomBytes(16).toString('hex')
-      const accessToken = await this.getMainAccessToken() || undefined
-      // Start HTTP server to handle OAuth callback
-      await this.oauthHttpServer.startServer((callbackData: OAuthCallbackData) => {
-        this.handleServerOAuthHttpCallback(callbackData, serverId, provider, { oauthToken: accessToken })
-      })
-
-      // Fetch authorization URL from server
-      const { authUrl, sessionId } = await this.oauthProviderManager.fetchServerAuthorizationUrl(
-        serverId,
-        provider,
-        state,
-        accessToken || undefined,
-      )
-
-      // Store session info for callback
-      this.currentProviderPKCE = {
-        codeVerifier: '',
-        codeChallenge: '',
-        state: state,
-        providerId: provider, // Use just the provider name (e.g., "google")
-        sessionId: sessionId,
-      }
-
-      // Open browser for user authentication
-      await shell.openExternal(authUrl)
-    }
-    catch (error) {
-      console.error(`❌ Server OAuth flow error for ${serverId}/${provider}:`, error)
-      this.notifyProviderAuthError(provider, 'Failed to start authentication')
-      this.oauthHttpServer.stopServer() // Clean up on error
-    }
-  }
-
-  /**
-   * Fetch onboarding GitHub provider
-   */
-  async fetchOnboardingGithubProvider(): Promise<void> {
-    const provider = 'onboarding'
-
-    const response = await fetch(`https://api.keyboard.dev/auth/keyboard_github/onboarding`)
-    const data = await response.json() as OnboardingGitHubResponse
-    const sessionId = data.session_id
-    const authUrl = data.authorization_url
-    const state = data.state
-    this.currentProviderPKCE = {
-      codeVerifier: '',
-      codeChallenge: '',
-      state: state,
-      providerId: provider, // Use just the provider name (e.g., "google")
-      sessionId: sessionId,
-    }
-    await this.oauthHttpServer.startServer((callbackData: OAuthCallbackData) => {
-      this.handleServerOAuthHttpCallback(callbackData, 'onboarding', provider)
-    })
-    if (!authUrl) throw new Error('No authorization URL found')
-    await shell.openExternal(authUrl)
-  }
-
-  /**
-   * Handle server OAuth HTTP callback
-   */
-  private async handleServerOAuthHttpCallback(
-    callbackData: OAuthCallbackData,
-    serverId: string,
-    provider: string,
-    options?: {
-      oauthToken?: string
-    },
-  ): Promise<void> {
-    try {
-      if (callbackData.error) {
-        throw new Error(`OAuth error: ${callbackData.error} - ${callbackData.error_description || ''}`)
-      }
-
-      if (!callbackData.code || !callbackData.state) {
-        throw new Error('Missing authorization code or state')
-      }
-
-      if (!this.currentProviderPKCE) {
-        throw new Error('No session data stored - possible callback timeout')
-      }
-
-      if (callbackData.state !== this.currentProviderPKCE.state) {
-        console.error('❌ State mismatch:', {
-          received: callbackData.state,
-          expected: this.currentProviderPKCE.state,
-        })
-        throw new Error('State mismatch - potential security issue')
-      }
-
-      // Exchange code for tokens using server
-      let accessToken = options?.oauthToken
-      if (!accessToken) {
-        accessToken = await this.getMainAccessToken() || undefined
-      }
-      const tokens = await this.oauthProviderManager.exchangeServerCodeForTokens(
-        serverId,
-        provider,
-        callbackData.code,
-        callbackData.state,
-        this.currentProviderPKCE.sessionId!,
-        accessToken || undefined,
-      )
-      // Store tokens securely
-      await this.perProviderTokenStorage.storeTokens(tokens)
-      if (provider === 'onboarding') {
-        await this.perProviderTokenStorage.saveOnboardingTokens(tokens)
-        const githubService = this.getGithubService()
-        await githubService.initializeToken()
-        await githubService.createFork('keyboard-dev', 'codespace-executor')
-        await githubService.createFork('keyboard-dev', 'app-creator')
-
-        // Connect to executor with the new onboarding token
-        const executorWSClient = this.getExecutorWSClient()
-        if (executorWSClient) {
-          executorWSClient.setGitHubToken(tokens.access_token)
-        }
-      }
-
-      this.currentProviderPKCE = null
-
-      // Notify the renderer process
-      const providerConfig = await this.oauthProviderManager.getProvider(provider)
-      this.windowManager.sendMessage('provider-auth-success', {
-        providerId: tokens.providerId,
-        providerName: providerConfig?.name || provider,
-        user: tokens.user,
-        authenticated: true,
-      })
-
-      // Show the window after successful authentication
-      this.windowManager.showWindow()
-
-      // Show success notification
-      this.showNotification({
-        id: `auth-success-${tokens.providerId}`,
-        title: `Server OAuth Authentication Successful`,
-        body: `Successfully connected to ${serverId} (${provider})${tokens.user ? ` as ${tokens.user.name || tokens.user.email}` : ''}`,
-        timestamp: Date.now(),
-        priority: 'normal',
-      })
-    }
-    catch (error) {
-      console.error(`❌ Server OAuth callback error for ${serverId}/${provider}:`, error)
-      this.notifyProviderAuthError(provider, `Authentication failed: ${error instanceof Error ? error.message : 'Unknown error'}`)
-    }
-  }
-
-  /**
-   * Get available OAuth providers
-   */
-  async getAvailableProviders(): Promise<OAuthProvider[]> {
-    return await this.oauthProviderManager.getAvailableProviders()
-  }
-
-  /**
-   * Get provider auth status
-   */
-  async getProviderAuthStatus(): Promise<Record<string, unknown>> {
-    return await this.perProviderTokenStorage.getProviderStatus()
-  }
-
-  /**
-   * Get provider tokens
-   */
-  async getProviderTokens(providerId: string): Promise<StoredProviderTokens | null> {
-    return await this.perProviderTokenStorage.getTokens(providerId)
-  }
-
-  /**
-   * Clear all provider tokens
-   */
-  async clearAllProviderTokens(): Promise<void> {
-    await this.perProviderTokenStorage.clearAllTokens()
-  }
-
-  /**
-   * Get OAuth storage info
-   */
-  getOAuthStorageInfo(): Record<string, unknown> {
-    return {
-      ...this.perProviderTokenStorage.getStorageInfo(),
-      providerStorage: this.oauthProviderManager.getProviderStorageInfo(),
-    }
-  }
-
-  /**
-   * Get all provider configs
-   */
-  async getAllProviderConfigs(): Promise<OAuthProviderConfig[]> {
-    return await this.oauthProviderManager.getAllProviderConfigs()
-  }
-
-  /**
-   * Save provider config
-   */
-  async saveProviderConfig(config: Omit<OAuthProviderConfig, 'createdAt' | 'updatedAt'>): Promise<void> {
-    await this.oauthProviderManager.saveProviderConfig(config)
-  }
-
-  /**
-   * Remove provider config
-   */
-  async removeProviderConfig(providerId: string): Promise<void> {
-    await this.oauthProviderManager.removeProviderConfig(providerId)
-  }
-
-  /**
-   * Get provider config
-   */
-  async getProviderConfig(providerId: string): Promise<OAuthProviderConfig | null> {
-    const provider = await this.oauthProviderManager.getProvider(providerId)
-    if (!provider) return null
-
-    // Get full config including metadata
-    const configs = await this.oauthProviderManager.getAllProviderConfigs()
-    return configs.find(c => c.id === providerId) || null
-  }
-
-  /**
-   * Add server provider
-   */
-  async addServerProvider(server: ServerProvider): Promise<void> {
-    await this.oauthProviderManager.addServerProvider(server)
-  }
-
-  /**
-   * Remove server provider
-   */
-  async removeServerProvider(serverId: string): Promise<void> {
-    await this.oauthProviderManager.removeServerProvider(serverId)
-  }
-
-  /**
-   * Get server providers
-   */
-  async getServerProviders(): Promise<ServerProvider[]> {
-    return await this.oauthProviderManager.getServerProviders()
-  }
-
-  /**
-   * Check if onboarding token exists
-   */
-  async checkOnboardingTokenExists(): Promise<boolean> {
-    return await this.perProviderTokenStorage.checkOnboardingTokenExists()
-  }
-
-  /**
-   * Clear onboarding token
-   */
-  async clearOnboardingToken(): Promise<void> {
-    await this.perProviderTokenStorage.clearOnboardingToken()
-  }
-
-  /**
-   * Fetch server providers
-   */
-  async fetchServerProviders(serverId: string): Promise<ServerProviderInfo[]> {
-    const accessToken = await this.getMainAccessToken()
-    const serverProviders = await this.oauthProviderManager.fetchServerProviders(serverId, accessToken || undefined)
-    return serverProviders
-  }
-
-  /**
-   * Handle WebSocket provider token request (for legacy WebSocket server)
-   */
-  async handleWebSocketProviderTokenRequest(providerId: string): Promise<{
-    type: string
-    providerId: string
-    token: string | null
-    encrypted: boolean
-    encryptionMethod: string
-    timestamp: number
-    authenticated: boolean
-    user?: unknown
-    providerName?: string
-    error?: string
-    requestId?: string
-  }> {
-    try {
-      const token = await this.getValidProviderAccessToken(providerId.toLowerCase())
-      const providerStatus = await this.perProviderTokenStorage.getProviderStatus()
-      const providerInfo = providerStatus[providerId]
-      const provider = await this.oauthProviderManager.getProvider(providerId)
-
-      // Encrypt token using codespace encryption if available
-      const { encryptedToken, encrypted, encryptionMethod } = token
-        ? await this.encryptProviderToken(token)
-        : { encryptedToken: token, encrypted: false, encryptionMethod: 'none' }
-
-      return {
-        type: 'provider-auth-token',
-        providerId: providerId,
-        token: encryptedToken,
-        encrypted: encrypted,
-        encryptionMethod: encryptionMethod,
-        timestamp: Date.now(),
-        authenticated: !!token || this.SKIP_AUTH,
-        user: providerInfo?.user || (this.SKIP_AUTH ? { email: 'test@example.com', firstName: 'Test Provider' } : null),
-        providerName: provider?.name || providerId,
-      }
-    }
-    catch (error) {
-      return {
-        type: 'provider-auth-token',
-        providerId: providerId,
-        token: null,
-        encrypted: false,
-        encryptionMethod: 'none',
-        error: `Failed to get token: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        timestamp: Date.now(),
-        authenticated: false,
-      }
-    }
-  }
-
-  /**
-   * Handle WebSocket provider status request (for legacy WebSocket server)
-   */
-  async handleWebSocketProviderStatusRequest(): Promise<{
-    type: string
-    tokensAvailable: string[]
-    timestamp: number
-    error?: string
-    requestId?: string
-  }> {
-    try {
-      const providerStatus = await this.perProviderTokenStorage.getProviderStatus()
-
-      // Check ALL stored provider tokens (both direct and server provider tokens)
-      const tokensAvailable = Object.entries(providerStatus)
-        .filter(([, status]) => status?.authenticated)
-        .map(([providerId]) => `KEYBOARD_PROVIDER_USER_TOKEN_FOR_${providerId.toUpperCase()}`)
-
-      return {
-        type: 'user-tokens-available',
-        tokensAvailable: tokensAvailable,
-        timestamp: Date.now(),
-      }
-    }
-    catch (error) {
-      return {
-        type: 'user-tokens-available',
-        tokensAvailable: [],
-        error: `Failed to get provider status: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        timestamp: Date.now(),
-      }
-    }
-  }
-
-  /**
-   * Stop OAuth HTTP server
-   */
-  stopOAuthHttpServer(): void {
-    this.oauthHttpServer.stopServer()
-  }
-
-  /**
-   * Expire all OAuth tokens for testing auto-refresh functionality
-   * This is a developer utility for testing
-   */
-  async expireAllTokensForTesting(): Promise<number> {
-    console.log('🧪 Expiring all OAuth tokens for testing...')
-    const count = await this.perProviderTokenStorage.expireAllTokensForTesting()
-    console.log(`🧪 Expired ${count} OAuth provider token(s)`)
-    return count
-  }
-}
